Move ClientLogFileEventArgs into its own file.
